refactor(webpack): move manifest stats logic to stats module
